% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cal.margin1}
\alias{cal.margin1}
\title{(Internal function) Calculate the first margin for random numbers, internally used for
parallelization}
\usage{
cal.margin1(prob, fst)
}
\arguments{
\item{prob}{A number to specify probability}

\item{fst}{A number to specify Fst}
}
\value{
The first margin for random numbers
}
\description{
(Internal function) Calculate the first margin for random numbers, internally used for
parallelization
}
