% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cal.prob.AA}
\alias{cal.prob.AA}
\title{(Internal function) Calculate probability for wild-type allele, internally used for
parallelization}
\usage{
cal.prob.AA(prob)
}
\arguments{
\item{prob}{A number to specify probability}
}
\value{
A probability of wild-type allele
}
\description{
(Internal function) Calculate probability for wild-type allele, internally used for
parallelization
}
