% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cal.prob.AB}
\alias{cal.prob.AB}
\title{(Internal) Calculate probability for heterozygous allele, internally used for
parallelization}
\usage{
cal.prob.AB(prob, riskratio = 1)
}
\arguments{
\item{prob}{A number to specify probability}

\item{riskratio}{A number to specify risk ratio for cases}
}
\value{
A probability of heterozygous allele
}
\description{
(Internal) Calculate probability for heterozygous allele, internally used for
parallelization
}
