% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cal.prob.BB}
\alias{cal.prob.BB}
\title{(Internal) Calculate probability for mutant-type allele, internally used for
parallelization}
\usage{
cal.prob.BB(prob, riskratio = 1)
}
\arguments{
\item{prob}{A number to specify probability}

\item{riskratio}{A number to specify risk ratio for cases}
}
\value{
A probability of mutant-type allele
}
\description{
(Internal) Calculate probability for mutant-type allele, internally used for
parallelization
}
