% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cal.prob.categorical.data}
\alias{cal.prob.categorical.data}
\title{(Internal) Calculate probabilities for categorical data, internally used for
parallelization}
\usage{
cal.prob.categorical.data(prob, max.category)
}
\arguments{
\item{prob}{A number to specify probability}

\item{max.category}{A number to specify the maximum number of categorical groups}
}
\value{
Probability for categorical data
}
\description{
(Internal) Calculate probabilities for categorical data, internally used for
parallelization
}
