% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{cbind.matrix}
\alias{cbind.matrix}
\title{(Internal) Combind two matrices by column for big data, internally used for
parallelization}
\usage{
\method{cbind}{matrix}(a, b)
}
\arguments{
\item{a}{The first matrix}

\item{b}{The second matrix}
}
\value{
The combined matrix by column
}
\description{
(Internal) Combind two matrices by column for big data, internally used for
parallelization
}
