% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{create.outlier}
\alias{create.outlier}
\title{(Internal) Create outliers,internally used for
parallelization}
\usage{
create.outlier(X, population, outlier)
}
\arguments{
\item{X}{An input matrix containing SNPs in additive coding}

\item{population}{A vector of labels for all individuals}

\item{outlier}{A number to specify the amount of outliers to be generated}
}
\value{
The modified matrix with outliers
}
\description{
(Internal) Create outliers,internally used for
parallelization
}
