% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{create.template.setting}
\alias{create.template.setting}
\title{Create a template for a setting file of function filest}
\usage{
create.template.setting(out.file, no.setting = 1)
}
\arguments{
\item{out.file}{An absolute path to a new setting file}

\item{no.setting}{A number of simulated settings}
}
\value{
An output directory if suggessfully created. Null if a setting file
can't be created.
}
\description{
Create a template for a setting file of function filest
}
\examples{

#Create 2 simulated settings

output <- file.path(tempdir(),"example_setting.txt")
res <- create.template.setting(out.file = output, no.setting = 2)
print(res)

}
