% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{demo.filest}
\alias{demo.filest}
\title{Demonstration the filest function}
\usage{
demo.filest()
}
\value{
The output directory
}
\description{
This function generates the setting file and demonstrate how to use \code{\link{filest}}.
}
\examples{

#To run this function, simply call demo.filest()
demo.filest()

}
