% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{do.create.cate}
\alias{do.create.cate}
\title{(Internal) Create categorical data, internally used for
parallelization}
\usage{
do.create.cate(marker, name)
}
\arguments{
\item{marker}{A string to specify a marker}

\item{name}{A string to specify a name of marker}
}
\value{
A categorical value
}
\description{
(Internal) Create categorical data, internally used for
parallelization
}
