% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{do.sample.categorical.data}
\alias{do.sample.categorical.data}
\title{(Internal) Sample categorical data,internally used for
parallelization}
\usage{
do.sample.categorical.data(prob, max.category, no.ind)
}
\arguments{
\item{prob}{A number to specify probability}

\item{max.category}{A number to specify the maximum number of categorical groups}

\item{no.ind}{A number to specify the amount of individuals to be generated}
}
\value{
A simulated data in categorical mode
}
\description{
(Internal) Sample categorical data,internally used for
parallelization
}
