% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{do.sample.SNP}
\alias{do.sample.SNP}
\title{(Internal function) Sample SNP according to given probabilities, internally used for
parallelization}
\usage{
do.sample.SNP(p.AA, p.AB, p.BB, no.ind)
}
\arguments{
\item{p.AA}{A number to specify probability of wild-type allele}

\item{p.AB}{A number to specify probability of heterozygous allele}

\item{p.BB}{A number to specify probability of mutant-type allele}

\item{no.ind}{A number to specify the amount of individuals to be generated}
}
\value{
A simulated SNP in additive coding
}
\description{
(Internal function) Sample SNP according to given probabilities, internally used for
parallelization
}
