% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FILEST_package.R
\docType{package}
\name{FILEST-package}
\alias{FILEST}
\alias{FILEST-package}
\title{Fine-Level Structure Simulator}
\description{
A population genetic simulator, which is able to generate synthetic datasets
for single-nucleotide polymorphisms (SNP) for multiple populations. The
genetic distances among populations can be set according to the Fixation
Index (Fst). This tool is able to simulate outlying individuals and missing
SNPs can be specified. For Genome-wide association study (GWAS), disease
status can be set in desired level according risk ratio.
}
\details{
The R package \pkg{FILEST} requires \pkg{KRIS} and \pkg{rARPACK}.

Here is the list of functions in the R package \pkg{FILEST}:
\itemize{
\item \code{\link{cbind_bigmatrix}}
\item \code{\link{create.template.setting}}
\item \code{\link{demo.filest}}
\item \code{\link{filest}}
\item \code{\link{rbind_bigmatrix}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/kris.ccp/filest}
  \item Report bugs at \url{https://gitlab.com/kris.ccp/filest/issues}
}

}
\author{
\strong{Maintainer}: Kridsadakorn Chaichoompu \email{kridsadakorn@biostatgen.org}

Authors:
\itemize{
  \item Kristel Van Steen
  \item Fentaw Abegaz
}

}
\keyword{internal}
