% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{generate.categorical.data}
\alias{generate.categorical.data}
\title{(Internal) Generate categorical data, internally used for
parallelization}
\usage{
generate.categorical.data(pop, fst, no.marker)
}
\arguments{
\item{pop}{A number to specify probability}

\item{fst}{A number to specify Fst}

\item{no.marker}{A number to specify the amount of markers to be generated}
}
\value{
A marrix of categorical data
}
\description{
(Internal) Generate categorical data, internally used for
parallelization
}
