% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{generate.label}
\alias{generate.label}
\title{(Internal) Generate labels, internally used for
parallelization}
\usage{
generate.label(pop, outlier)
}
\arguments{
\item{pop}{A number to specify the amount of populations}

\item{outlier}{A number to specify the amount of outliers}
}
\value{
A vector of population labels
}
\description{
(Internal) Generate labels, internally used for
parallelization
}
