% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{generate.snp}
\alias{generate.snp}
\title{(Internal function) Generate SNPs, internally used for
parallelization}
\usage{
generate.snp(pop, fst, no.snp, riskratio = 1)
}
\arguments{
\item{pop}{A number to specify the amount of populations}

\item{fst}{A number to specify Fst (representing genetic distance)}

\item{no.snp}{A number to specify the amount of SNPs}

\item{riskratio}{A number to specify risk ratio for cases}
}
\value{
Simulated SNPs
}
\description{
(Internal function) Generate SNPs, internally used for
parallelization
}
