% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{get.random.beta}
\alias{get.random.beta}
\title{(Internal function) Get random number based on beta distribution, internally used for
parallelization}
\usage{
get.random.beta(mar1, mar2)
}
\arguments{
\item{mar1}{A number to specify the first margin for random function of beta
distribution.}

\item{mar2}{A number to specify the second margin for random function of beta
distribution.}
}
\value{
A random number based on beta distribution
}
\description{
(Internal function) Get random number based on beta distribution, internally used for
parallelization
}
