% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{rbind.matrix}
\alias{rbind.matrix}
\title{(Internal) Combind two matrices by row for big data, internally used for
parallelization}
\usage{
\method{rbind}{matrix}(a, b)
}
\arguments{
\item{a}{The first matrix}

\item{b}{The second matrix}
}
\value{
The combined matrix by row
}
\description{
(Internal) Combind two matrices by row for big data, internally used for
parallelization
}
