% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{rbind_bigmatrix}
\alias{rbind_bigmatrix}
\title{Combind two matrices by row for big data, internally used for
parallelization}
\usage{
rbind_bigmatrix(a, b)
}
\arguments{
\item{a}{The first matrix}

\item{b}{The second matrix}
}
\value{
The combined matrix by row
}
\description{
Combind two matrices by row for big data, internally used for
parallelization
}
\examples{
X <- matrix(c(1,2,0,1,2,2,1,2,0,0,1,2,1,2,2,2),ncol=4)
Y <- matrix(c(2,1,1,0,1,0,0,1,1,2,2,0,0,1,1,0),ncol=4)
Z <- rbind_bigmatrix(X,Y)
print(Z)
}
\seealso{
\code{\link{cbind_bigmatrix}}
}
