% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{save.PC.plot}
\alias{save.PC.plot}
\title{(Internal) Generate PC plots labels, internally used for
parallelization}
\usage{
save.PC.plot(file.name, PC, label)
}
\arguments{
\item{file.name}{A string to specify the output file name}

\item{PC}{A matrix of principal components (PC)}

\item{label}{A vector of string to specify population labels}
}
\description{
(Internal) Generate PC plots labels, internally used for
parallelization
}
