% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{filest}
\alias{filest}
\title{Simulate data for multiple populations}
\usage{
filest(setting, out, thread = 1)
}
\arguments{
\item{setting}{An absolute path to a setting file}

\item{out}{An absolute path for output files}

\item{thread}{A number to specify a maximum thread to be run in parallel}
}
\description{
The output files are saved to the specified directory according to \code{out}.
}
\examples{

#See the examples from
demo.filest()
}
