% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filest.R
\name{get.para}
\alias{get.para}
\title{(Internal function) Get all parameters}
\usage{
get.para(param)
}
\arguments{
\item{param}{A string of parameters}
}
\value{
A list of parameters
}
\description{
(Internal function) Get all parameters
}
